requirements.txt
python-dotenv==1.0.1
openai==1.54.3
fastapi==0.115.4
uvicorn==0.32.0
langchain==0.3.7
langchain_community==0.3.5